feature: improved transports
This adds a new NNG transport and improve how Server and Client uses the
transport.

NNG support multiple in-flight requests by using a nng.Context to make
sure the req-req state machine is respected. So server can effectively
do req, req and then later send each rep in the order they need.

This is important because our servers process multiple RPCs at a time,
bounded by our thread pool, and rpc calls will be answered in an
unpredictable order, so a normal req-rep cycle won't work.

This also adds a orjson-based JSON serializer.

Transport is now created by a factory that handles os compatibility.
redis transport is only support on non-win32 systems, so that factory
handles that. If you don't specify a transport protocol, it will use the
default for that platform. Redis is default on non-win32, NNG is default
on win32.